Separate scripts created
Separate scripts for the all the placements. (aka table placement, stove and trash). They where created to have a order so that when its on the table it can't go anywhere but the stove since its higher priority and when on the stove it can't return to the table. the trash will be the reset but hasn't been implemented yet.  once its implemented then adding things to the kettle to remember is the next step.