WIP
Current issue the placements don't work if you enter the are and don't leave it except for table. the stove and trash seem to not respond properly unless the kettle has left there are once. when the trash is used once it screws with the table placement leaving only the stove working for some reason. might try switching the collisions to squares otherwise might have to think of another system to make it work.